Azure ile Çeviri Uygulaması/requirements.txt
flask

python-dotenv

requests